config-loader: separate out source test utils
